% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cams_rad.r
\name{cams_get_mcclear}
\alias{cams_get_mcclear}
\title{Retrieve McClear clear sky solar radiation data from
  \href{http://www.soda-pro.com/web-services/radiation/cams-radiation-service}{CAMS
  radiation service}}
\usage{
cams_get_mcclear(username, lat, lng, date_begin, date_end, alt = -999,
  time_step = "PT01H", time_ref = "UT", verbose = FALSE)
}
\arguments{
\item{username}{The email registered at soda-pro.com.}

\item{lat}{The latitude}

\item{lng}{The longitude}

\item{date_begin}{Start date as 'yyy-mm-dd' string}

\item{date_end}{End date as 'yyy-mm-dd' string}

\item{alt}{The altitude in meters, use -999 to let CAMS decide}

\item{time_step}{'PT01M' for minutes, 'PT15M' for 15 minutes, 'PT01H' for
hourly, 'P01D' for daily, 'P01M' for monthly}

\item{time_ref}{'UT' for universal time, 'TST' for true solar time}

\item{verbose}{TRUE for verbose output}
}
\value{
A tibble (data.frame) with requested solar data
}
\description{
Retrieve McClear clear sky solar radiation data from
  \href{http://www.soda-pro.com/web-services/radiation/cams-radiation-service}{CAMS
  radiation service}
}
\examples{
\dontrun{
username <- "your@email.com" # An email registrated at soda-pro.com
df <- cams_get_mcclear(username, lat=60, lon=15,
                       date_begin="2016-01-01", date_end="2016-01-15")
print(df)
}

}

