% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cams_rad.r
\name{cams_api}
\alias{cams_api}
\title{API client for
  \href{http://www.soda-pro.com/web-services/radiation/cams-radiation-service}{CAMS
  radiation service}}
\usage{
cams_api(username, lat, lng, date_begin, date_end, alt = -999,
  time_step = "PT01H", time_ref = "UT", verbose = FALSE,
  service = "get_cams_radiation", format = "application/csv",
  filename = "")
}
\arguments{
\item{username}{The email registered at soda-pro.com.}

\item{lat}{The latitude}

\item{lng}{The longitude}

\item{date_begin}{Start date as 'yyy-mm-dd' string}

\item{date_end}{End date as 'yyy-mm-dd' string}

\item{alt}{The altitude in meters, use -999 to let CAMS decide}

\item{time_step}{'PT01M' for minutes, 'PT15M' for 15 minutes, 'PT01H' for
hourly, 'P01D' for daily, 'P01M' for monthly}

\item{time_ref}{'UT' for universal time, 'TST' for true solar time}

\item{verbose}{TRUE for verbose output}

\item{service}{'get_mcclear' for CAMS McClear data, get_cams_radiation for CAMS
radiation data}

\item{format}{'application/csv', 'application/json',
'application/x-netcdf' or 'text/csv'}

\item{filename}{path to file on disk to write to. If empty, data is kept in memory.}
}
\value{
list(ok=TRUE/FALSE, response=response). If ok=TRUE, response is the
  response from httr::GET. If ok=FALSE, response holds exception text
}
\description{
API client for
  \href{http://www.soda-pro.com/web-services/radiation/cams-radiation-service}{CAMS
  radiation service}
}
\examples{
\dontrun{
library(ncdf4)

filename <- paste0(tempfile(), ".nc")

r <- cams_api(username, 60, 15, "2016-06-01", "2016-06-10",
              format = "application/x-netcdf", filename=filename)

# Access the on disk stored ncdf4 file
nc <- nc_open(r$respone$content)
# list names of available variables
names(nc$var)

# create data.frame with datetime and global horizontal irradiation
df <- data.frame(datetime=as.POSIXct(nc$dim$time$vals, "UTC",
                                     origin="1970-01-01"),
                 GHI = ncvar_get(nc, "GHI"))

plot(df, type="l")
}

}

