% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cams_get.R
\name{cams_get_radiation}
\alias{cams_get_radiation}
\title{Retrieve CAMS solar radiation data}
\usage{
cams_get_radiation(lat, lng, date_begin, date_end, time_step = "PT01H",
  alt = -999, verbose = FALSE)
}
\arguments{
\item{lat}{The latitude}

\item{lng}{The longitude}

\item{date_begin}{Start date as 'yyy-mm-dd' string}

\item{date_end}{End date as 'yyy-mm-dd' string}

\item{time_step}{'PT01M' for minutes, 'PT15M' for 15 minutes, 'PT01H' for
hourly, 'P01D' for daily, 'P01M' for monthly}

\item{alt}{The altitude in meters, use -999 to let CAMS decide}

\item{verbose}{TRUE for verbose output}

\item{time_ref}{'UT' for universal time, 'TST' for true solar time}

\item{service}{'get_mcclear' for CAMS McClear data, get_cams_radiation for CAMS
radiation data}

\item{format}{'application/csv', 'application/json',
'application/x-netcdf' or 'text/csv'}
}
\value{
A data frame with requested solar data
}
\description{
Retrieve CAMS solar radiation data
}
\examples{
\dontrun{
df <- cams_get_radiation(
  lat=60, lng=15, date_begin="2016-01-01", date_end="2016-01-15")
print(head(df))
}

}

